fix(重新整理proto文件):

+++ mxshop-api/user-web/response/user.go
@@ -13,7 +13,7 @@ func (j JsonTime) MarshalJSON() ([]byte, error) {
 }
 
 type UserResponse struct {
-	Id       int32  `json:"id"`
+	Id       uint64 `json:"id"`
 	NickName string `json:"nickName"`
 	//Birthday time.Time `json:"birthday"`
 	Birthday JsonTime `json:"birthday"`
